pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["Daniil Nikulin <d.nikulin@innopolis.university>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.45.0"
extra-streamlit-components = "^0.1.80"
pytest = "^8.3.5"
bandit = "^1.8.3"
flake8 = "^7.2.0"
mccabe = "^0.7.0"
pytest-cov = "^6.1.1"
flake8-max-function-length = "^0.9.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
